pyproject.toml
[project]
name = "mcp-x"
version = "0.1.0"
description = "MCP X project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.1.2",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "tweepy>=4.14.0",
]

[[project.authors]]
name = "Daoyi Lu"
email = "lyons2000124@outlook.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-x = "mcp_x:main"
